Generate interaction data set

Simulate a single data set with an interaction (y ~ x1 + x2 + x1*x2). All values other than 'N' are population-level effects - the values within any single simulated data set will vary around the defined values.

generate_interaction(
  N,
  r.x1.y,
  r.x2.y,
  r.x1x2.y,
  r.x1.x2,
  rel.x1 = 1,
  rel.x2 = 1,
  rel.y = 1,
  skew.x1 = 0,
  skew.x2 = 0,
  skew.y = 0,
  transform.x1 = "default",
  transform.x2 = "default",
  transform.y = "default",
  adjust.correlations = T,
  r.x1.y.adjust = NULL,
  r.x2.y.adjust = NULL,
  r.x1x2.y.adjust = NULL,
  r.x1.x2.adjust = NULL
)

Arguments

N
Sample size. Must be a positive integer. Has no default value.

r.x1.y
Pearson's correlation between x1 and y. Must be between -1 and 1. Has no default value.

r.x2.y
Pearson's correlation between x2 and y. Must be between -1 and 1. Assumed to be the 'moderator' in some functions. Has no default value.

r.x1x2.y
Pearson's correlation between the interaction term x1x2 (x1 * x2) and y. Must be between -1 and 1. Has no default value.

r.x1.x2
Pearson's correlation between x1 and x2. Must be between -1 and 1. Has no default value.

rel.x1
Reliability of x1 (e.g. test-retest reliability, ICC, Cronbach's alpha). Default is 1 (perfect reliability). Must be greater than 0 and less than or equal to 1.

rel.x2
Reliability of x2 (e.g. test-retest reliability, ICC, Cronbach's alpha). Default is 1 (perfect reliability). Must be greater than 0 and less than or equal to 1.

rel.y
Reliability of xy (e.g. test-retest reliability, ICC, Cronbach's alpha). Default is 1 (perfect reliability). Must be greater than 0 and less than or equal to 1.

skew.x1
Skew of x1. Default is 0 (normally distributed).

skew.x2
Skew of x2. Default is 0 (normally distributed).

skew.y
Skew of y. Default is 0 (normally distributed).

transform.x1
Transform x1? Options are "default", "binary", or "gamma". "binary" will cause variable to be binarized - 2 unique values. Default ("default") will pick "gamma" if variable is skewed.

transform.x2
Transform x2? Options are "default", "binary", or "gamma". "binary" will cause variable to be binarized - 2 unique values. Default ("default") will pick "gamma" if variable is skewed.

transform.y
Transform y? Options are "default", "binary", or "gamma". "binary" will cause variable to be binarized - 2 unique values. Default ("default") will pick "gamma"if variable is skewed.

adjust.correlations
If variables are skewed or binary, should correlations be adjusted so that output data has the specified correlation structure? Default is TRUE.

r.x1.y.adjust
Internal use only

r.x2.y.adjust
Internal use only

r.x1x2.y.adjust
Internal use only

r.x1.x2.adjust
Internal use only

Value

A data frame containing variables 'x1', 'x2', 'y', and 'x1x2'. 'x1x2' is x1*x2. The correlations between these variables are drawn from the defined population-level values.

Examples

if (FALSE) dataset <- generate_interaction(N = 250,r.x1.y = 0,r.x2.y = .1,r.x1x2.y = -.2,r.x1.x2 = .3)
cor(dataset)

#> Error in is.data.frame(x): object 'dataset' not found