Test interaction

Test the interaction from a single simulated data set.

Usage

test_interaction(data, alpha = 0.05, q = 2, simple = FALSE)

Arguments

data

Simulated data set. Output of 'generate_interaction()'.

alpha

The alpha. At what p-value is the interaction deemed significant? Default is 0.05.

q

Simple slopes. How many quantiles should x2 be split into for simple slope testing? Default is 2. Simple slope testing returns the effect-size (slope) of y~x1 for the two most extreme quantiles of x2. If q=3 then the two slopes are y~x1 for the bottom 33% of x2, and the top 33% of x2.

simple

For internal use. Default is FALSE.

Value

Either a named list or a data frame containing the results of the regression y~x1+x2+x1*x2, the pearson's correlation between y, x1,x2, and x1x2, and the slopes of the simple slopes.

Examples

dataset <- generate_interaction(N = 250,r.x1.y = 0,r.x2.y = .1,r.x1x2.y = -.2,r.x1.x2 = .3)
test_interaction(data = dataset, alpha=0.05, q=2)
#> $linear.model
#>         Estimate Std. Error   t value    Pr(>|t|)
#> x1   -0.07121674 0.06513502 -1.093371 0.275300429
#> x2    0.17468645 0.06566831  2.660133 0.008324440
#> x1x2 -0.15599605 0.05690505 -2.741339 0.006568058
#> 
#> $x1x2.adjusted.r2
#> [1] 0.02487781
#> 
#> $x1x2.confint
#>   x1x2_95confint_2.5 x1x2_95confint_97.5
#> 1         -0.2680793         -0.04391278
#> 
#> $interaction.shape
#>   crossover.point    shape
#> 1        1.119813 2.190441
#> 
#> $simple.slopes
#>    lower.slope upper.slope
#> x2   0.1654821  -0.2278128
#> 
#> $correlation
#>           cor v1   v2
#> 5   0.3000000 x1   x2
#> 9   0.0157794 x1    y
#> 10  0.1942301 x2    y
#> 13 -0.1974594 x1 x1x2
#> 14 -0.2335283 x2 x1x2
#> 15 -0.2019081  y x1x2
#>